hiring_flow_prediction/requirements.txt
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
et-xmlfile==1.1.0
fonttools==4.42.0
joblib==1.3.2
kiwisolver==1.4.4
matplotlib==3.7.2
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.23.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.21.12
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.3.0

